items can now be returned (fixing ticket #8)
git-svn-id: https://shell.noname-ev.de/svn/kasse/c128@59 af93e077-1a23-4f1e-9cbe-9382a9d578f5

+++ src/kasse.c
@@ -135,7 +135,7 @@ void buy(BYTE n) {
 		 * to NULL if no such credit could be found */
 		for (c = 0; c < credits.num_items; ++c)
 			if (strncmp(nickname, credits.credits[c].nickname, nickname_len) == 0) {
-				if (credits.credits[c].credit < (status.status[n].price * einheiten)) {
+				if ((signed int)credits.credits[c].credit < ((signed int)status.status[n].price * einheiten)) {
 					cprintf("Sorry, %s hat nicht genug Geld :-(\r\n", nickname);
 					return;
 				}
